Started CentOS deployment
Before: I was hoping to deploy on Fedora, but unfortunately, I ran into
dependency issues in regards to mysqllibs.

Now: I am planning to deploy to CentOS 7 since it is an offically
supported distro in the docs. I also moved the repo installation into
its own role and I started working on the kvm-node role.